fix: 5주차 수정(AuthContext 파일 분리)
