+++ week06/mission/Chap6_mission_01/src/context/AuthContext.tsx
@@ -1,9 +1,5 @@
-import { LOCAL_STORAGE_KEY } from "../constants/key";
-import { useLocalStorage } from "../hooks/useLocalStorage";
 import type { RequestSigninDto } from "../types/auth";
-import React, { createContext, type PropsWithChildren } from "react";
-import { useState, useContext } from "react";
-import { postSignin, postLogout } from "../apis/auth";
+import React, { createContext } from "react";
 
 interface AuthContextType {
   accessToken: string | null;
@@ -19,75 +15,3 @@ export const AuthContext: React.Context<AuthContextType> =
     login: async () => {},
     logout: async () => {},
   });
-
-export const AuthProvider = ({ children }: PropsWithChildren) => {
-  const {
-    getItem: getAccessTokenFromStorage,
-    setItem: setAccessTokenFromStorage,
-    removeItem: removeAccessTokenFromStorage,
-  } = useLocalStorage(LOCAL_STORAGE_KEY.accessToken);
-
-  const {
-    getItem: getRefreshTokenFromStorage,
-    setItem: setRefreshTokenFromStorage,
-    removeItem: removeRefreshTokenFromStorage,
-  } = useLocalStorage(LOCAL_STORAGE_KEY.refreshToken);
-
-  const [accessToken, setAccessToken] = useState<string | null>(
-    getAccessTokenFromStorage()
-  );
-
-  const [refreshToken, setRefreshToken] = useState<string | null>(
-    getRefreshTokenFromStorage()
-  );
-
-  const login = async (signInData: RequestSigninDto) => {
-    try {
-      const { data } = await postSignin(signInData);
-      if (data) {
-        const newAccessToken = data.accessToken;
-        const newRefreshToken = data.refreshToken;
-
-        setAccessTokenFromStorage(newAccessToken);
-        setRefreshTokenFromStorage(newRefreshToken);
-
-        setAccessToken(newAccessToken);
-        setRefreshToken(newRefreshToken);
-        alert("로그인 성공");
-        // window.location.href = "/";
-      }
-    } catch (e) {
-      console.error("로그인 오류", e);
-      alert("로그인 실패");
-    }
-  };
-
-  const logout = async () => {
-    try {
-      await postLogout();
-      removeAccessTokenFromStorage();
-      removeRefreshTokenFromStorage();
-      setAccessToken(null);
-      setRefreshToken(null);
-      alert("로그아웃 성공");
-    } catch (e) {
-      console.error("로그아웃 오류", e);
-      alert("로그아웃 실패");
-    }
-  };
-
-  return (
-    <AuthContext.Provider value={{ accessToken, refreshToken, login, logout }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-
-  if (!context) {
-    throw new Error("AuthContext를 찾을 수 없습니다.");
-  }
-  return context;
-};
