+++ week06/mission/Chap6_mission_01/src/context/AuthProvider.tsx
@@ -1,0 +1,70 @@
+import { LOCAL_STORAGE_KEY } from "../constants/key";
+import { useLocalStorage } from "../hooks/useLocalStorage";
+import type { RequestSigninDto } from "../types/auth";
+import { type PropsWithChildren } from "react";
+import { useState } from "react";
+import { postSignin, postLogout } from "../apis/auth";
+import { AuthContext } from "./AuthContext";
+
+export const AuthProvider = ({ children }: PropsWithChildren) => {
+  const {
+    getItem: getAccessTokenFromStorage,
+    setItem: setAccessTokenFromStorage,
+    removeItem: removeAccessTokenFromStorage,
+  } = useLocalStorage(LOCAL_STORAGE_KEY.accessToken);
+
+  const {
+    getItem: getRefreshTokenFromStorage,
+    setItem: setRefreshTokenFromStorage,
+    removeItem: removeRefreshTokenFromStorage,
+  } = useLocalStorage(LOCAL_STORAGE_KEY.refreshToken);
+
+  const [accessToken, setAccessToken] = useState<string | null>(
+    getAccessTokenFromStorage()
+  );
+
+  const [refreshToken, setRefreshToken] = useState<string | null>(
+    getRefreshTokenFromStorage()
+  );
+
+  const login = async (signInData: RequestSigninDto) => {
+    try {
+      const { data } = await postSignin(signInData);
+      if (data) {
+        const newAccessToken = data.accessToken;
+        const newRefreshToken = data.refreshToken;
+
+        setAccessTokenFromStorage(newAccessToken);
+        setRefreshTokenFromStorage(newRefreshToken);
+
+        setAccessToken(newAccessToken);
+        setRefreshToken(newRefreshToken);
+        alert("로그인 성공");
+        // window.location.href = "/";
+      }
+    } catch (e) {
+      console.error("로그인 오류", e);
+      alert("로그인 실패");
+    }
+  };
+
+  const logout = async () => {
+    try {
+      await postLogout();
+      removeAccessTokenFromStorage();
+      removeRefreshTokenFromStorage();
+      setAccessToken(null);
+      setRefreshToken(null);
+      alert("로그아웃 성공");
+    } catch (e) {
+      console.error("로그아웃 오류", e);
+      alert("로그아웃 실패");
+    }
+  };
+
+  return (
+    <AuthContext.Provider value={{ accessToken, refreshToken, login, logout }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
