+++ week06/mission/Chap6_mission_01/src/hooks/useAuth.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+import { AuthContext } from "../context/AuthContext";
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+
+  if (!context) {
+    throw new Error("AuthContext를 찾을 수 없습니다.");
+  }
+  return context;
+};
