+++ week06/mission/Chap6_mission_01/src/hooks/useForm.tsx
@@ -6,11 +6,10 @@ interface UseFormProps<T> {
   validate: (values: T) => Record<keyof T, string>;
 }
 
-function useForm<T>({initialValue, validate}: UseFormProps<T>) {
-  const [values, setValues] = useState(initialValue);
-  // 눌렀는지 안눌렀는지
-  const [touched, setTouched] = useState<Record<string, boolean>>(); 
-  const [errors, setErrors] = useState<Record<string, string>>();
+function useForm<T>({ initialValue, validate }: UseFormProps<T>) {
+  const [values, setValues] = useState(initialValue); // 폼 입력값
+  const [touched, setTouched] = useState<Record<string, boolean>>(); // 입력한적이 있는지
+  const [errors, setErrors] = useState<Record<string, string>>(); // validate 함수가 반환한 에러 메시지
 
   // 사용자가 입력값을 바꿀 때 실행
   const handleChange = (name: keyof T, text: string) => {
@@ -24,16 +23,17 @@ function useForm<T>({initialValue, validate}: UseFormProps<T>) {
     setTouched({
       ...touched,
       [name]: true,
-    })
-  }
+    });
+  };
 
-  // 이메일 인풋, 패스워드 인풋, 속성들을 좀 가져오는 것
+  // 이메일 인풋, 패스워드 인풋, 속성들을 가져오는 것
   const getInputProps = (name: keyof T) => {
     const value = values[name];
-    const onChange = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => handleChange(name, e.target.value);
+    const onChange = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) =>
+      handleChange(name, e.target.value);
     const onBlur = () => handleBlur(name);
 
-    return {value, onChange, onBlur};
+    return { value, onChange, onBlur };
   };
 
   // values가 변경될 때마다 에러 검증 로직 실행
@@ -42,7 +42,7 @@ function useForm<T>({initialValue, validate}: UseFormProps<T>) {
     setErrors(newErrors); // 오류 메시지 업데이트
   }, [validate, values]);
 
-  return {values, errors, touched, getInputProps};
+  return { values, errors, touched, getInputProps };
 }
 
 export default useForm;
