+++ week06/mission/Chap6_mission_01/src/layout/ProtectedLayout.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet, useLocation } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 import Navbar from "../components/Navbar";
 import { useEffect, useRef } from "react";
